+++ example/ios/InstabugTests/CrashReportingApiTests.m
@@ -4,12 +4,13 @@
 #import "Instabug/IBGCrashReporting.h"
 #import "Instabug/Instabug.h"
 #import "Util/Instabug+Test.h"
+#import "Util/IBGCrashReporting+CP.h"
 
 @interface CrashReportingApiTests : XCTestCase
 
-@property (nonatomic, strong) id mCrashReporting;
-@property (nonatomic, strong) id mInstabug;
-@property (nonatomic, strong) CrashReportingApi *api;
+@property(nonatomic, strong) id mCrashReporting;
+@property(nonatomic, strong) id mInstabug;
+@property(nonatomic, strong) CrashReportingApi *api;
 
 @end
 
@@ -24,9 +25,9 @@
 - (void)testSetEnabled {
     NSNumber *isEnabled = @1;
     FlutterError *error;
-
+    
     [self.api setEnabledIsEnabled:isEnabled error:&error];
-
+    
     OCMVerify([self.mCrashReporting setEnabled:YES]);
 }
 
@@ -34,10 +35,32 @@
     NSString *jsonCrash = @"{}";
     NSNumber *isHandled = @0;
     FlutterError *error;
-
+    
     [self.api sendJsonCrash:jsonCrash isHandled:isHandled error:&error];
+    
+    OCMVerify([self.mCrashReporting cp_reportFatalCrashWithStackTrace:@{}]);
+}
 
-    OCMVerify([self.mInstabug reportCrashWithStackTrace:@{} handled:isHandled]);
+
+- (void)testSendNonFatalErrorJsonCrash {
+    NSString *jsonCrash = @"{}";
+    NSString *fingerPrint = @"fingerprint";
+    NSDictionary *userAttributes = @{@"key": @"value",};
+    NSString *ibgNonFatalLevel = @"NonFatalExceptionLevel.error";
+    
+    FlutterError *error;
+    
+    [self.api sendNonFatalErrorJsonCrash:jsonCrash
+                          userAttributes:userAttributes
+                             fingerprint:fingerPrint
+                  nonFatalExceptionLevel:ibgNonFatalLevel
+                                   error:&error];
+    
+    OCMVerify([self.mCrashReporting cp_reportNonFatalCrashWithStackTrace:@{}
+                level:IBGNonFatalLevelError
+                groupingString:fingerPrint
+                userAttributes:userAttributes
+              ]);
 }
 
 @end